+++ lib/pages/practice/widgets/button_panel.dart
@@ -42,33 +42,17 @@ class ButtonPanel extends StatelessWidget {
         width: double.infinity,
         margin: const EdgeInsets.all(8),
         child: Row(
-          children: (manager.practiceMode == PracticeMode.reviewByDueDate) //
-              ? _buttonsWithDate()
-              : _plainButtons(),
-        ),
+            children: switch (manager.practiceMode) {
+          PracticeMode.reviewBySpacedRepetition => _spacedRepetitionButtons(),
+          PracticeMode.reviewByFixedDays => _fixedDaysButtons(),
+          PracticeMode.reviewSameNumberPerDay => _plainButtons(),
+          PracticeMode.casualPractice => _plainButtons(),
+        }),
       ),
     );
   }
 
-  List<Widget> _plainButtons() {
-    return [
-      Expanded(
-        child: ResponseButton(
-          title: 'Again',
-          onPressed: () => manager.onResponse(Difficulty.hard),
-        ),
-      ),
-      const SizedBox(width: 5),
-      Expanded(
-        child: ResponseButton(
-          title: 'Good',
-          onPressed: () => manager.onResponse(Difficulty.good),
-        ),
-      ),
-    ];
-  }
-
-  List<Widget> _buttonsWithDate() {
+  List<Widget> _spacedRepetitionButtons() {
     return [
       Expanded(
         child: ResponseButton(
@@ -85,6 +69,60 @@ class ButtonPanel extends StatelessWidget {
           subtitle: manager.goodTitle,
           onPressed: () => manager.onResponse(Difficulty.good),
           onLongPress: () => manager.onResponse(Difficulty.easy),
+        ),
+      ),
+    ];
+  }
+
+  List<Widget> _fixedDaysButtons() {
+    return [
+      Expanded(
+        child: ResponseButton(
+          title: 'Hard',
+          subtitle: manager.hardTitle,
+          onPressed: () => manager.onResponse(Difficulty.hard),
+        ),
+      ),
+      const SizedBox(width: 5),
+      Expanded(
+        child: ResponseButton(
+          title: 'OK',
+          subtitle: manager.okTitle,
+          onPressed: () => manager.onResponse(Difficulty.ok),
+        ),
+      ),
+      const SizedBox(width: 5),
+      Expanded(
+        child: ResponseButton(
+          title: 'Good',
+          subtitle: manager.goodTitle,
+          onPressed: () => manager.onResponse(Difficulty.good),
+        ),
+      ),
+      const SizedBox(width: 5),
+      Expanded(
+        child: ResponseButton(
+          title: 'Easy',
+          subtitle: manager.easyTitle,
+          onPressed: () => manager.onResponse(Difficulty.easy),
+        ),
+      ),
+    ];
+  }
+
+  List<Widget> _plainButtons() {
+    return [
+      Expanded(
+        child: ResponseButton(
+          title: 'Again',
+          onPressed: () => manager.onResponse(Difficulty.hard),
+        ),
+      ),
+      const SizedBox(width: 5),
+      Expanded(
+        child: ResponseButton(
+          title: 'Good',
+          onPressed: () => manager.onResponse(Difficulty.good),
         ),
       ),
     ];